requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
beautifulsoup4==4.10.0
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.4.1
Django==3.2.11
django-bootstrap4==21.2
django-easy-pdf==0.1.1
django-utils-six==2.0
django-wkhtmltopdf==3.3.0
fonttools==4.28.5
future==0.18.2
html5lib==1.1
importlib-metadata==2.1.2
Pillow==9.0.0
pycparser==2.21
pydyf==0.1.2
PyPDF2==1.26.0
pyphen==0.12.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.5
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tinycss2==1.1.1
typing_extensions==4.0.1
weasyprint==54.0
webencodings==0.5.1
xhtml2pdf==0.2.5
zipp==3.7.0
zopfli==0.1.9
